[FIX] It avoids validating the password_history when updating the auth_crypt module.
When installing the module project_forecast automatically updates the
auth_crypt module and this module has an init that updates all the
users passwords, but assigning the same password, which causes a
password error used, if the password_history is different Of 0. To avoid
this, use the value 'everything' that comes in the context when a module
is being updated.